backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
supabase==1.0.3
python-multipart==0.0.6
anthropic==0.5.0
together==0.2.5
cohere==4.32
huggingface-hub==0.19.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1 